+++ cmp/grid/GridModel.js
@@ -109,6 +109,8 @@ export class GridModel extends HoistModel {
     /** @member {ReactNode} */
     restoreDefaultsWarning;
     /** @member {boolean} */
+    fullRowEditing;
+    /** @member {boolean} */
     hideEmptyTextBeforeLoad;
     /** @member {boolean} */
     highlightRowOnClick;
@@ -242,6 +244,7 @@ export class GridModel extends HoistModel {
      *      obscured by horizontal scrolling. Note that setting this value to true may limit the
      *      ability of the grid to autosize offscreen columns effectively. Default false.
      * @param {GridAutosizeOptions} [c.autosizeOptions] - default autosize options.
+     * @param {boolean} [c.fullRowEditing] - true to enable full row editing. Default false.
      * @param {boolean} [c.externalSort] - Set to true to if application will be
      *      reloading data when the sortBy property changes on this model (either programmatically,
      *      or via user-click.)  Useful for applications with large data sets that are performing
@@ -305,6 +308,7 @@ export class GridModel extends HoistModel {
         useVirtualColumns = false,
         autosizeOptions = {},
         restoreDefaultsWarning = GridModel.DEFAULT_RESTORE_DEFAULTS_WARNING,
+        fullRowEditing = false,
         highlightRowOnClick = XH.isPhone || XH.isTablet,
         experimental,
         ...rest
@@ -335,6 +339,7 @@ export class GridModel extends HoistModel {
             fillMode: 'none'
         });
         this.restoreDefaultsWarning = restoreDefaultsWarning;
+        this.fullRowEditing = fullRowEditing;
         this.highlightRowOnClick = highlightRowOnClick;
 
         apiRemoved(rest.contextMenuFn, 'contextMenuFn', 'Use contextMenu instead');
@@ -1187,6 +1192,7 @@ export class GridModel extends HoistModel {
             const field = storeFields.find(f => f.name === col.field);
             if (!field) return col;
 
+            // TODO: Set the editor based on field type
             return {
                 displayName: field.displayName,
                 align: numTypes.includes(field.type) ? 'right' : undefined,