+++ desktop/cmp/grid/editors/index.js
@@ -4,9 +4,9 @@
  *
  * Copyright © 2021 Extremely Heavy Industries Inc.
  */
-export * from './Form';
-export * from './FormModel';
-export * from './field/BaseFieldModel';
-export * from './field/FieldModel';
-export * from './field/SubformsFieldModel';
-
+export * from './CheckboxEditor';
+export * from './DateEditor';
+export * from './NumberEditor';
+export * from './SelectEditor';
+export * from './TextEditor';
+export * from './TextAreaEditor';