+++ admin/tabs/activity/clienterrors/ClientErrorsModel.js
@@ -111,8 +111,8 @@ export class ClientErrorsModel extends HoistModel {
         });
 
         this.filterChooserModel = new FilterChooserModel({
-            sourceStore: this.gridModel.store,
-            targetStore: this.gridModel.store,
+            filterSource: this.gridModel.store,
+            filterTarget: this.gridModel.store,
             fieldSpecs: [
                 'username',
                 'browser',