+++ CHANGELOG.md
@@ -70,6 +70,7 @@
 * filesize `6.2 -> 6.3`
 * mobx `6.1 -> 6.3`
 * react-windowed-select `3.0 -> 3.1`
+* highcharts `8.2.2 -> 9.1.2`
 
 [Commit Log](https://github.com/xh/hoist-react/compare/v40.0.0...develop)